requirements.txt
cffi==1.16.0
cryptography==41.0.4
ecdsa==0.18.0
pycparser==2.21
pycryptodome==3.18.0
six==1.16.0
